requirements.txt
streamlit
tensorflow-cpu
numpy
requests
groq
Pillow
fpdf
streamlit-option-menu